program now populatd the TUI with live data from ESPN

+++ src/model.rs
@@ -1,9 +1,12 @@
-#[derive(Clone)]
+use serde::{Serialize, Deserialize};
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub enum League {
     Wbc,
     Mlb,
 }
 
+#[derive(Debug, Serialize, Deserialize)]
 pub enum GameStatus {
     Scheduled,
     InProgress,
@@ -13,12 +16,14 @@ pub enum GameStatus {
     // Delayed,
 }
 
+#[derive(Debug, Serialize, Deserialize)]
 pub struct BaseState {
     pub on_first: bool,
     pub on_second: bool,
     pub on_third: bool,
 }
 
+#[derive(Debug, Serialize, Deserialize)]
 pub struct GameSummary {
     pub game_id: String,
     pub league: League,
@@ -26,15 +31,15 @@ pub struct GameSummary {
     pub away_team_abbrev: String,
     pub home_team_abbrev: String,
 
-    pub away_team_score: Option<u8>,
-    pub home_team_score: Option<u8>,
+    pub away_team_score: Option<u64>,
+    pub home_team_score: Option<u64>,
 
     pub game_status: GameStatus,
     pub status_text: String,
 
-    pub balls: Option<u8>,
-    pub strikes: Option<u8>,
-    pub outs: Option<u8>,
+    pub balls: Option<u64>,
+    pub strikes: Option<u64>,
+    pub outs: Option<u64>,
 
     pub bases: Option<BaseState>,
 }
